re-structure flask app
