--- Backend/Apply/models.py
@@ -1,4 +1,0 @@
-#Defines database models and sets up SQLAlchemy or an ORM for database operations
-from flask_sqlalchemy import SQLAlchemy
-
-db = SQLAlchemy()
