--- Backend/apply/admin.py
@@ -1,5 +1,0 @@
-from app import app
-
-@app.route('/viewApplicants')
-def viewApplicants():
-    return 'View Applicants'
